Refactor chat and sidebar into modular components
