+++ components/chat/index.ts
@@ -1,0 +1,1 @@
+export { ChatShell as default } from './ChatShell';
