+++ components/sidebar/DeleteConfirmationModal.tsx
