+++ components/sidebar/index.ts
@@ -1,0 +1,4 @@
+export { SidebarShell as default } from './SidebarShell';
+export { ProjectSection } from './ProjectSection';
+export { ConversationMenu } from './ConversationMenu';
+export { ConversationListItem } from './ConversationListItem';
